Added old command line functionality back in, renamed deployable files to server.html and server.js. Also added .envexample for the deploy.sh to work from.
