+++ .env.example
@@ -1,0 +1,2 @@
+PROJECT_ID=
+SERVICE_NAME=
